Bug Fix: Positions from xml unavailable when template style is not default
Unable to choose positions when template style is not set as default. In
2 places the database call to determine the template name which was in 1
place used to determine the path to the templateDetails.xml, was based
on if a template style helix3 was set as the default. This created
unwanted dysfunction during testing of different templates that was not
yet used or set as default. This has now fixed by determining the
template name based on the ID in the database of the template style that
you are currently editing.